refactor: 갤러리 페이지 도메인 재정립
