--- src/entities/gallery/GalleryFilter.ts
@@ -1,4 +1,0 @@
-export interface GalleryFilterType {
-  setSelectedCategory: React.Dispatch<React.SetStateAction<string>>;
-  selectedCategory: string;
-}
