--- src/entities/gallery/GalleryItem.ts
@@ -1,13 +1,0 @@
-export interface GalleryItem {
-  id?: string;
-  title: string;
-  description: string;
-  imageUrl: string;
-  category: string;
-  photographer: string;
-  eventDate: string;
-  views: number;
-  likes: number;
-  createdAt: string;
-  updatedAt: string;
-}
